pyproject.toml
[project]
name = "django-allauth-sms"
version = "0.1.0"
dependencies = [
    "django-allauth >= 0.52.0",
    "django-sms >= 0.6",
    "django-phonenumber-field[phonenumbers] >= 7.0.2",
]
